more reorganisation (make some "namespace"-like modules more consistent with dir structure
