+++ spec/contexts/errors.rb
@@ -1,5 +1,3 @@
-require './spec/mock/reader'
-
 RSpec.shared_context 'errors' do
   include_context 'lexer'
